refactor: update component directory
