+++ src/components/Error/styled.tsx
@@ -1,0 +1,20 @@
+import styled from 'styled-components'
+
+export const ErrorPanel = styled.div`
+  margin: 112px 0 203px 0;
+
+  @media (max-width: 750px) {
+    margin: 120px 0 130px 0;
+  }
+
+  > img {
+    width: 1038px;
+    height: 480px;
+    margin: 0 auto;
+
+    @media (max-width: 750px) {
+      width: 282px;
+      height: 130px;
+    }
+  }
+`
