+++ src/components/Header/MobileMenu/styled.tsx
@@ -1,0 +1,16 @@
+import styled from 'styled-components'
+
+export const MobileMenusPanel = styled.div`
+  width: 100%;
+  height: 100%;
+  background: #1c1c1c;
+  display: flex;
+  flex-direction: column;
+  position: fixed;
+  position: -webkit-fixed;
+  z-index: 2;
+  color: white;
+  top: 64px;
+  bottom: 0px;
+  overflow: hidden;
+`
