+++ src/components/Modal/styled.tsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components'
+
+export const ModalPanel = styled.div`
+  display: block;
+  position: fixed;
+  z-index: 1;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  overflow: auto;
+  background-color: rgba(0, 0, 0, 0.4);
+`
