+++ src/components/Search/UDTSearch/index.tsx
@@ -1,73 +1,13 @@
 import React, { useState, useRef, useEffect, useMemo } from 'react'
-import styled from 'styled-components'
 import { useTranslation } from 'react-i18next'
-import SearchLogo from '../../assets/search_black.png'
-import ClearLogo from '../../assets/clear.png'
-import i18n from '../../utils/i18n'
-import { AppActions, ComponentActions } from '../../contexts/providers/reducer'
-import { useDispatch, useAppState } from '../../contexts/providers'
-import { getSimpleUDTTransactionsWithAddress } from '../../service/app/udt'
-import { isMobile } from '../../utils/screen'
-
-const SearchPanel = styled.div`
-  width: 360px;
-  height: 38px;
-  text-align: center;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-
-  @media (max-width: 1440px) {
-    width: 320px;
-  }
-
-  @media (max-width: 750px) {
-    width: 84vw;
-  }
-`
-
-const SearchImage = styled.div`
-  display: inline-block;
-  margin-right: -36px;
-  margin-left: 8px;
-  width: 18px;
-  z-index: 2;
-  display: flex;
-  justify-content: center;
-
-  @media (max-width: 750px) {
-    margin-right: ${(props: { isRight?: boolean }) => (props.isRight ? '8px' : '-36px')};
-    margin-left: ${(props: { isRight?: boolean }) => (props.isRight ? '-24px' : '8px')};
-  }
-
-  img {
-    width: 18px;
-    height: 18px;
-  }
-`
-
-const SearchInputPanel = styled.input`
-  position: relative;
-  width: 100%;
-  height: 100%;
-  font-size: 16px;
-  padding-left: ${(props: { isRight?: boolean }) => (props.isRight ? '8px' : '30px')};
-  padding-right: ${(props: { isRight?: boolean }) => (props.isRight ? '30px' : '8px')};
-  background: rgba(255, 255, 255, 0);
-  opacity: 1;
-  border: 1px solid #b3b3b3;
-  color: #666666;
-  border-radius: 6px;
-
-  &: focus {
-    color: #666666;
-    outline: none;
-  }
-
-  &::placeholder {
-    color: #bababa;
-  }
-`
+import SearchLogo from '../../../assets/search_black.png'
+import ClearLogo from '../../../assets/clear.png'
+import i18n from '../../../utils/i18n'
+import { AppActions, ComponentActions } from '../../../contexts/providers/reducer'
+import { useDispatch, useAppState } from '../../../contexts/providers'
+import { getSimpleUDTTransactionsWithAddress } from '../../../service/app/udt'
+import { isMobile } from '../../../utils/screen'
+import { UDTSearchInputPanel, UDTSearchImage, UDTSearchPanel } from './styled'
 
 const clearSearchInput = (inputElement: any) => {
   const input: HTMLInputElement = inputElement.current
@@ -140,7 +80,7 @@ const UDTSearch = ({ typeHash, content }: { typeHash: string; content?: string }
   const ClearIconButton = () => {
     const dispatch = useDispatch()
     return (
-      <SearchImage
+      <UDTSearchImage
         isRight={true}
         role="button"
         tabIndex={-1}
@@ -155,14 +95,14 @@ const UDTSearch = ({ typeHash, content }: { typeHash: string; content?: string }
         }}
       >
         <img src={ClearLogo} alt="search clear logo" />
-      </SearchImage>
+      </UDTSearchImage>
     )
   }
 
   return (
-    <SearchPanel>
+    <UDTSearchPanel>
       {!searchBarEditable && (
-        <SearchImage
+        <UDTSearchImage
           role="button"
           tabIndex={-1}
           onKeyPress={() => {}}
@@ -171,9 +111,9 @@ const UDTSearch = ({ typeHash, content }: { typeHash: string; content?: string }
           }}
         >
           <img src={SearchLogo} alt="search logo" />
-        </SearchImage>
+        </UDTSearchImage>
       )}
-      <SearchInputPanel
+      <UDTSearchInputPanel
         isRight={isMobile() && searchBarEditable}
         ref={inputElement}
         placeholder={placeholder}
@@ -191,7 +131,7 @@ const UDTSearch = ({ typeHash, content }: { typeHash: string; content?: string }
         }}
       />
       {searchBarEditable && <ClearIconButton />}
-    </SearchPanel>
+    </UDTSearchPanel>
   )
 }
