+++ src/components/Text/index.tsx
@@ -1,30 +1,8 @@
 import React from 'react'
-import styled from 'styled-components'
 import { Link } from 'react-router-dom'
 import { Tooltip } from 'antd'
 import CopyTooltipText from './CopyTooltipText'
-
-export const HighLightPanel = styled.div`
-  color: ${props => props.theme.primary};
-  font-size: 14px;
-
-  @media (max-width: 750px) {
-    font-size: 13px;
-  }
-
-  a {
-    color: ${props => props.theme.primary};
-    margin-top: 3px;
-
-    @media (max-width: 750px) {
-      margin-top: 1px;
-    }
-  }
-
-  a:hover {
-    color: ${props => props.theme.primary};
-  }
-`
+import { HighLightPanel } from './styled'
 
 export const HighLightLink = ({ value, to, tooltip }: { value: string; to: string; tooltip?: string }) => {
   return tooltip ? (
