+++ src/components/Toast/index.tsx
@@ -1,54 +1,7 @@
 import React, { useState, useEffect, useReducer } from 'react'
-import styled from 'styled-components'
 import { useTimeoutWithUnmount } from '../../utils/hook'
 import { useAppState } from '../../contexts/providers'
-
-const ToastDiv = styled.div`
-  position: absolute;
-  position: -webkit-absolute;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  box-sizing: border-box;
-  display: flex;
-  z-index: 9998;
-  flex-direction: column;
-  pointer-events: none;
-`
-const ToastItemDiv = styled.div`
-  width: 100%;
-  position: fixed;
-  position: -webkit-fixed;
-  top: 64px;
-  opacity: 0.96;
-  z-index: 9999;
-  height: 60px;
-  .toast__text {
-    color: white;
-    font-size: 20px;
-    line-height: 60px;
-    text-align: center;
-  }
-
-  @media (max-width: 750px) {
-    top: 42px;
-    height: 36px;
-
-    .toast__text {
-      font-size: 14px;
-      line-height: 36px;
-    }
-  }
-
-  @media (max-width: 320px) {
-    top: 42px;
-    height: 36px;
-    .toast__text {
-      font-size: 12px;
-      line-height: 36px;
-    }
-  }
-`
+import { ToastItemPanel, ToastPanel } from './styled'
 
 const getColor = (type: 'success' | 'warning' | 'danger') => {
   switch (type) {
@@ -95,14 +48,14 @@ const ToastItem = ({ data, willLeave }: { data: State.ToastMessage; willLeave: F
   )
 
   return (
-    <ToastItemDiv
+    <ToastItemPanel
       style={{
         opacity,
         background: getColor(data.type),
       }}
     >
       <div className="toast__text">{data.message}</div>
-    </ToastItemDiv>
+    </ToastItemPanel>
   )
 }
 
@@ -146,7 +99,7 @@ export default () => {
   }, [dispatch, app.toast])
 
   return state.toasts.length === 0 ? null : (
-    <ToastDiv className="toast">
+    <ToastPanel className="toast">
       {state.toasts &&
         state.toasts.map((item: State.ToastMessage) => {
           return (
@@ -164,6 +117,6 @@ export default () => {
             />
           )
         })}
-    </ToastDiv>
+    </ToastPanel>
   )
 }
