+++ src/pages/TransactionList/index.tsx
@@ -4,7 +4,8 @@ import queryString from 'query-string'
 import { useTranslation } from 'react-i18next'
 import { parseSimpleDate } from '../../utils/date'
 import Content from '../../components/Content'
-import { TableTitleRow, TableTitleItem, TableContentRow, TableContentItem } from '../../components/Table'
+import { TableTitleRow, TableContentRow } from '../../components/Table/styled'
+import { TableTitleItem, TableContentItem } from '../../components/Table'
 import { shannonToCkb } from '../../utils/util'
 import { parsePageNumber, adaptMobileEllipsis, adaptPCEllipsis } from '../../utils/string'
 import { ListPageParams } from '../../utils/const'
